+++ Sources/SwiftUIKit/Views/Flip/FlipDirection.swift
@@ -1,0 +1,17 @@
+//
+//  FlipDirection.swift
+//  SwiftUIKit
+//
+//  Created by Daniel Saidi on 2020-01-05.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This enum defines flip directions to use with ``FlipView``.
+ */
+public enum FlipDirection {
+    
+    case left, right, up, down
+}